requirements.txt
# RAG Course Planning Assistant - Requirements
# Python 3.10+

# Core LLM & RAG
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.41
langchain-openai==0.1.23

# Vector Store
faiss-cpu==1.8.0

# Embeddings (HuggingFace - free alternative)
sentence-transformers==3.1.1
torch==2.4.1
transformers==4.45.2

# OpenAI (optional - for GPT-4 based generation)
openai==1.51.0

# Anthropic Claude (alternative LLM backend)
anthropic==0.34.2

# Document Processing
beautifulsoup4==4.12.3
requests==2.32.3
html2text==2024.2.26

# Data handling
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
tqdm==4.66.5
tiktoken==0.7.0

# Visualization
matplotlib==3.9.2
networkx==3.3
graphviz==0.20.3

# Web UI
streamlit==1.39.0
gradio==4.44.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# API
fastapi==0.115.2
uvicorn==0.31.0

# Logging & Monitoring
loguru==0.7.2
